Improves README, Makefile and adds -h flag.
The README was incredibly bare. This commit adds much more details to it,
including dependencies, the matrix format and how to use and build the programs.

The Makefile had a lot of unneeded junk which I have now removed. I also added a
rule to ensure that the build/ directory was created if it didn't exist. Clean
will also now remove both executables.

The qvdraw script can also now print out a short usage decription using the `-h`
flag.